Fixed circular dependency schema
