+++ packages/core/src/schemas/api.ts
@@ -15,6 +15,8 @@ import {
 } from "./schedules";
 import { CachedTaskSchema, TaskSchema } from "./tasks";
 import { EventSpecificationSchema, TriggerMetadataSchema } from "./triggers";
+import { RunStatusSchema } from "./runs";
+import { JobRunStatusRecord } from "./statuses";
 
 export const UpdateTriggerSourceBodyV1Schema = z.object({
   registeredEvents: z.array(z.string()),
@@ -730,32 +732,7 @@ export const CreateExternalConnectionBodySchema = z.object({
 
 export type CreateExternalConnectionBody = z.infer<typeof CreateExternalConnectionBodySchema>;
 
-const StatusUpdateStateSchema = z.union([
-  z.literal("loading"),
-  z.literal("success"),
-  z.literal("failure"),
-]);
-export type StatusUpdateState = z.infer<typeof StatusUpdateStateSchema>;
-
-const StatusUpdateDataSchema = z.record(SerializableJsonSchema);
-export type StatusUpdateData = z.infer<typeof StatusUpdateDataSchema>;
-
-export const StatusUpdateSchema = z.object({
-  label: z.string().optional(),
-  state: StatusUpdateStateSchema.optional(),
-  data: StatusUpdateDataSchema.optional(),
+export const GetRunStatusesSchema = z.object({
+  run: z.object({ id: z.string(), status: RunStatusSchema }),
+  statuses: z.array(JobRunStatusRecord),
 });
-export type StatusUpdate = z.infer<typeof StatusUpdateSchema>;
-
-const InitalStatusUpdateSchema = StatusUpdateSchema.required({ label: true });
-export type InitialStatusUpdate = z.infer<typeof InitalStatusUpdateSchema>;
-
-export const StatusHistorySchema = z.array(StatusUpdateSchema);
-export type StatusHistory = z.infer<typeof StatusHistorySchema>;
-
-export const JobRunStatusRecord = InitalStatusUpdateSchema.extend({
-  key: z.string(),
-  history: StatusHistorySchema,
-});
-
-export const GetRunStatusesSchema = z.object({ statuses: z.array(JobRunStatusRecord) });
