+++ packages/core/src/schemas/index.ts
@@ -12,3 +12,4 @@ export * from "./fetch";
 export * from "./events";
 export * from "./runs";
 export * from "./addMissingVersionField";
+export * from "./statuses";
